+++ Editor/EditorSections.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd76f3c76a82c074086b167673f6619a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
