+++ Editor/EditorSections/SelectTagsSection.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 174051d780c405d4c88ff0054fab7d01
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
